requirements.txt
flask==3.0.3
groq==0.4.2
gtts==2.5.0
python-dotenv==1.0.1
requests==2.31.0
